Move Scenario to the core.test package and make it public
