+++ test/org/javarosa/xpath/expr/XPathPathExprCurrentGroupCountRefTest.java
@@ -20,6 +20,7 @@ import static org.hamcrest.Matchers.is;
 import static org.javarosa.xpath.expr.AnswerDataMatchers.stringAnswer;
 import static org.junit.Assert.assertThat;
 
+import org.javarosa.core.test.Scenario;
 import org.junit.Before;
 import org.junit.Test;
