+++ test/org/javarosa/xpath/expr/XPathPathExprCurrentItemsetNodesetTest.java
@@ -19,6 +19,7 @@ import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.javarosa.xpath.expr.SelectChoiceMatchers.choice;
 import static org.junit.Assert.assertThat;
 
+import org.javarosa.core.test.Scenario;
 import org.junit.Before;
 import org.junit.Test;
